add DeleteModalWithConfirm component
